packages/hldpro-sim/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "hldpro-sim"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pydantic>=2.0", "PyYAML>=6.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["hldprosim*"]

[tool.setuptools.package-data]
hldprosim = [
  "package_data/personas/*.json",
  "package_data/scholar/*.md",
  "package_data/scholar/*.yaml",
  "package_data/scholar/schemas/*.json"
]
